requirements.txt
# Fetch.ai uAgents
uagents==0.12.0
cosmpy==0.9.2

# LLM & AI
openai==1.12.0
langchain==0.1.9
groq==0.4.2

# API & Web
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0
aiohttp==3.9.3

# Monitoring & Data
psutil==5.9.8
prometheus-client==0.19.0

# Web Scraping (Bright Data)
beautifulsoup4==4.12.3
selenium==4.18.1

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
redis==5.0.1
loguru==0.7.2

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
